# src/docs/swagger.yaml
basePath: /
definitions:
...
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
  model.Faq:
    properties:
      active:
        type: boolean
      answer:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      question:
        type: string
      removedAt:
        type: string
    type: object
...
  model.VendorApplication:
    properties:
      approvedAt:
        type: string
      aptNumber:
        type: string
      birthday:
        type: string
      city:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      notes:
        type: string
      phoneNumber:
        type: string
      referralCode:
        type: string
      rejectedAt:
        type: string
      socialLink1:
        type: string
      socialLink2:
        type: string
      socialLink3:
        type: string
      state:
        type: string
      status:
        type: string
      streetAddress:
        type: string
      submittedAt:
        type: string
      uid:
        type: string
      username:
        type: string
      zipCode:
        type: string
    type: object
...
host: apidev.xopacks.com
info:
  contact: {}
  description: The official XOPAcks API
  title: XOPacks API
  version: "1.0"
paths:
  /admin/addFaq:
    post:
      consumes:
      - application/json
      description: Adds an faq to the list of active faq
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      - description: faq
        in: body
        name: faq
        required: true
        schema:
          $ref: '#/definitions/model.Faq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Faq'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Add an faq
      tags:
      - Admin
  /admin/approveVendor:
    post:
      consumes:
      - application/json
      description: Approve a vendors application and verify their account
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      - description: creator uid
        in: query
        name: vendorUid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.VendorApplication'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Approve a new vendors application
      tags:
      - Admin
  /admin/editFaq:
    patch:
      consumes:
      - application/json
      description: Changes the structure of an existing faq
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      - description: faq id
        in: query
        name: id
        required: true
        type: integer
      - description: faq patch map
        in: body
        name: faqPatch
        required: true
        schema:
          $ref: '#/definitions/model.Faq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Edit an existing faq
      tags:
      - Admin
  /admin/login:
    post:
      consumes:
      - application/json
      description: Authenticate and login to the admin dashboard
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Login as an admin
      tags:
      - Admin
  /admin/rejectVendor:
    post:
      consumes:
      - application/json
      description: Reject a vendorss application
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      - description: creator uid
        in: query
        name: vendorUid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.VendorApplication'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Reject a vendor application
      tags:
      - Admin
  /admin/removeCreator:
    delete:
      consumes:
      - application/json
      description: Revoke a vendors verification
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      - description: creator uid
        in: query
        name: vendorUid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Remove a vendor
      tags:
      - Admin
  /admin/removeFaq:
    delete:
      consumes:
      - application/json
      description: Remoe an faq from the active list
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      - description: faq id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Remove an faq
      tags:
      - Admin
...
swagger: "2.0"